use float

+++ include/teleop_manager/teleop_manager.hpp
@@ -30,7 +30,7 @@ class TeleopManager : public rclcpp::Node
         float max_x_velocity_;
         float max_y_velocity_;
         float max_yawrate_;
-        double max_auto_velocity_;
+        float max_auto_velocity_;
 
         //flags
         bool stop_flag_ = 0;
